+++ morph/Matrix22.h
@@ -74,26 +74,26 @@ namespace morph {
         }
 
         //! Access elements of the matrix
-        Flt& operator[] (size_t idx) { return this->mat[idx]; }
+        Flt& operator[] (unsigned int idx) { return this->mat[idx]; }
         // note: assume Flt is a built-in type here (safe - Flt will be float or double)
-        const Flt operator[] (size_t idx) const  { return this->mat[idx]; }
+        const Flt operator[] (unsigned int idx) const  { return this->mat[idx]; }
 
         //! Access a given row of the matrix
-        morph::vec<Flt, 2> row (size_t idx) const
+        morph::vec<Flt, 2> row (unsigned int idx) const // unsigned char would be enough capacity!
         {
             morph::vec<Flt, 2> r = {Flt{0}, Flt{0}};
-            if (idx > 1) { return r; }
+            if (idx > 1U) { return r; }
             r[0] = this->mat[idx];
             r[1] = this->mat[idx+2];
             return r;
         }
 
         //! Access a given column of the matrix
-        morph::vec<Flt, 2> col (size_t idx) const
+        morph::vec<Flt, 2> col (unsigned int idx) const
         {
             morph::vec<Flt, 3> c = {Flt{0}, Flt{0}};
-            if (idx > 1) { return c; }
-            idx *= 2;
+            if (idx > 1U) { return c; }
+            idx *= 2U;
             c[0] = this->mat[idx];
             c[1] = this->mat[++idx];
             return c;
